Make datasource implementation internal
