+++ app/src/test/java/gq/kirmanak/mealient/ui/baseurl/BaseURLViewModelTest.kt
@@ -5,7 +5,7 @@ import gq.kirmanak.mealient.data.auth.AuthRepo
 import gq.kirmanak.mealient.data.baseurl.ServerInfoRepo
 import gq.kirmanak.mealient.data.recipes.RecipeRepo
 import gq.kirmanak.mealient.datasource.NetworkError
-import gq.kirmanak.mealient.datasource.impl.TrustedCertificatesStoreImpl
+import gq.kirmanak.mealient.datasource.TrustedCertificatesStore
 import gq.kirmanak.mealient.test.AuthImplTestData.TEST_BASE_URL
 import gq.kirmanak.mealient.test.BaseUnitTest
 import gq.kirmanak.mealient.ui.OperationUiState
@@ -33,7 +33,7 @@ class BaseURLViewModelTest : BaseUnitTest() {
     lateinit var recipeRepo: RecipeRepo
 
     @MockK(relaxUnitFun = true)
-    lateinit var trustedCertificatesStore: TrustedCertificatesStoreImpl
+    lateinit var trustedCertificatesStore: TrustedCertificatesStore
 
     lateinit var subject: BaseURLViewModel
