+++ datasource/src/main/kotlin/gq/kirmanak/mealient/datasource/DataSourceModule.kt
@@ -31,7 +31,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-interface DataSourceModule {
+internal interface DataSourceModule {
 
     companion object {
